refactor: Area 관련 DTO 패키지 분리 및 불필요한 클래스 삭제 (#2)
