--- src/main/java/com/chocobi/leafy/place/dto/AreaApiResponse.java
@@ -1,11 +1,0 @@
-package com.chocobi.leafy.place.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class AreaApiResponse {
-    private int code;
-    private String name;
-}
