--- src/main/java/com/chocobi/leafy/place/dto/AreaCodeItem.java
@@ -1,9 +1,0 @@
-package com.chocobi.leafy.place.dto;
-
-import lombok.Data;
-
-@Data
-public class AreaCodeItem {
-    private String code;
-    private String name;
-}
