--- src/main/java/com/chocobi/leafy/place/dto/AreaCodeResponse.java
@@ -1,8 +1,0 @@
-package com.chocobi.leafy.place.dto;
-
-import lombok.Data;
-
-@Data
-public class AreaCodeResponse<T> {
-    private Response<T> response;
-}
