+++ src/main/java/com/chocobi/leafy/place/dto/area/AreaCodeApiResponse.java
@@ -1,0 +1,9 @@
+package com.chocobi.leafy.place.dto.area;
+
+import com.chocobi.leafy.place.dto.common.Response;
+import lombok.Data;
+
+@Data
+public class AreaCodeApiResponse<T> {
+    private Response<T> response;
+}
