+++ src/main/java/com/chocobi/leafy/place/dto/area/AreaCodeItem.java
@@ -1,0 +1,13 @@
+package com.chocobi.leafy.place.dto.area;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class AreaCodeItem {
+    private String code;
+    private String name;
+}
